fix: move event_key_utils functionality into main file
